Revert "[libcxx][pretty printers] Import gdb module in gdb feature check"
This reverts commit 0c2a4548455c6c943ac5e2b5c51ed5c2c91e34e7.

This was my mistake. When gdb can find its data directory it'll
import it automatically. If it can't (like when you're using a
version from a build folder) you need to give it the data
directory path.

We're safe to assume gdb is installed for testing purposes
so it'll import it for us.

NOKEYCHECK=True
GitOrigin-RevId: 5efafc3e65c52effeefa84385bc9d8b7eb0d3291